added new module bubbleChart along with new lib
1. The floating bar button now is used by the bubble chart button. An ID
is assigned. (Only the FB group is called Canvas Team Mandarine, the
official name is just Team "Mandarine");
2.  bubbleChart.js is a separate module which for now creates a chart
with static data. An addon for ChartJS is used.
Known issues: slow framerate.
3. Chart.Scatter.js is a new library, an addon for Chart.js It works
only if you have Chart.js file connected with html file. It is not
minified. There is a minified version.
4. Added tvBubbleChart variable, corrected a typo in "To be editted",
added new case for the ID of the buble-chart button.

Known issues: the first three buttons when pressed prevents the use of
the radar and bubble modules.